chore(.editorconfig): add .editorconfig file to enforce consistent coding style feat(build.gradle.kts): add function to load gradle.properties file and set properties as system properties fix(Dependencies.kt): change usage of System.getenv to System.getProperty to retrieve KEYCLOAK_VERSION
The .editorconfig file is added to the project to enforce consistent coding style across different editors and IDEs. It specifies the indentation size and disabled rules for wildcard imports.

In the build.gradle.kts file, a function `loadGradleProperties()` is added to load the gradle.properties file and set the properties as system properties. This allows the properties to be accessed within the build script.

In the Dependencies.kt file, the usage of `System.getenv("KEYCLOAK_VERSION")` is changed to `System.getProperty("KEYCLOAK_VERSION")` to retrieve the value of the KEYCLOAK_VERSION property. This change ensures that the property is retrieved correctly.