chore(api-config): update Keycloak client credentials in application.yml chore(im-script-gateway): remove unnecessary auth configuration in application.yml chore(im-script-gateway): update Keycloak client credentials in application.yml chore(.env_dev): update VERSION_IM to 'local' chore(dev-compose.mk): add step to generate config_root.json file using envsubst feat(docker-compose-im-root-config.yml): add docker-compose file for im-root-config service
The Keycloak client credentials in the application.yml file for the api-config and im-script-gateway services have been updated to use new values. The unnecessary auth configuration in the application.yml file for the im-script-gateway service has been removed. The VERSION_IM environment variable in the .env_dev file has been updated to 'local'. The dev-compose.mk file has been modified to include a step to generate the config_root.json file using envsubst. The docker-compose-im-root-config.yml file has been added to include the im-root-config service configuration. These changes are made to update the configuration and improve the functionality of the services.