ci(dev.yml): update PKG_GITHUB_USERNAME to use github.actor for better clarity ci(dev.yml): update PKG_GITHUB_TOKEN to use secrets.GITHUB_TOKEN for consistency ci(sec.yml): remove PKG_GITHUB_USERNAME and PKG_GITHUB_TOKEN as they are no longer needed ci(sec.yml): add permissions for packages read access for improved security
The changes in the dev.yml workflow update the PKG_GITHUB_USERNAME to use github.actor for better clarity and PKG_GITHUB_TOKEN to use secrets.GITHUB_TOKEN for consistency in handling secrets. In the sec.yml workflow, the PKG_GITHUB_USERNAME and PKG_GITHUB_TOKEN are removed as they are no longer needed. Additionally, permissions for packages read access are added for improved security.