feat(UserPage.feature): add scenario to query users by givenName in UserPage feature
The scenario "I want to get a page of users filtered by givenName" is added to the UserPage feature file. This scenario allows users to query a page of users based on the givenName attribute. By introducing this scenario, the feature becomes more comprehensive and covers a common use case of filtering users by a specific attribute.