ci: remove .gitlab-ci.yml to discontinue GitLab CI integration
The .gitlab-ci.yml file is deleted to discontinue the use of GitLab CI/CD for this project. This change might be due to a shift to another CI/CD platform or a decision to manage CI/CD processes differently. Removing this file ensures that the GitLab CI/CD pipeline is no longer triggered.