style(login.ftl): update placeholder text to remove "pseudo" for clarity The placeholder text in the login form is updated to remove "pseudo" and only display "Adresse Email" to simplify the user interface and reduce confusion.
style(login.css): reorganize and update CSS for login page
The CSS is updated to improve the styling of the login page. The changes include adding underlines to links, updating button styles, and enhancing form control appearance. This improves the visual consistency and user experience of the login page.

refactor(template.ftl): comment out username display block
The block displaying the username is commented out because the username is an UUID, which is not user-friendly. This change prevents displaying non-intuitive information to the user, enhancing the login experience.