chore(env): add KC_REALM_ROOT_CLIENT_SECRET to .env_dev for keycloak configuration
Add the `KC_REALM_ROOT_CLIENT_SECRET` variable to the `.env_dev` file to
support secure authentication for the Keycloak realm root client. This
enhancement is necessary to ensure that the application can securely
interact with Keycloak, especially in development environments.

chore: remove notion.md file

Remove the `notion.md` file as it is no longer needed. The file contained
outdated or unnecessary information related to Notion API requests and
data, which is not relevant to the current project scope. This cleanup
helps in maintaining a more organized and relevant codebase.